+++ tests/functional/pytorch/017_aitune_torch_tune_resnet_tensorrt_backend_with_onnx_dynamo_test.py
@@ -37,7 +37,7 @@ def do_test(backend: TensorRTBackend, dtype: torch.dtype, device: str):
     expected_probs = torch.nn.functional.softmax(out[0], dim=0)
 
     strategy = OneBackendStrategy(backend)
-    strategy.enable_validate_against_baseline(False)
+    strategy.enable_performance_validation(False)
     module = Module(model, "functional-resnet18", strategy=strategy)
 
     # when