+++ tests/functional/pytorch/024_aitune_torch_toy_model_tensorrt_backend.py
@@ -48,7 +48,7 @@ def testing_multi_profile_with_samples():
     # configure the backend to use multi-profile mode, and auto generate profiles from samples used for tuning
     backend = TensorRTBackend(TensorRTBackendConfig(profiles=ProfileMode.SAMPLES_USED))
     strategy = OneBackendStrategy(backend)
-    strategy.enable_validate_against_baseline(False)
+    strategy.enable_performance_validation(False)
     strategy.enable_find_max_batch_size(False)
     module = Module(
         model,